requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorclass==2.2.2
docopt==0.6.2
EasyProcess==1.1
entrypoint2==1.1
google-ai-generativelanguage==0.6.9
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.66.0
gTTS==2.5.3
httplib2==0.22.0
idna==3.8
mss==9.0.1
packaging==24.1
pillow==10.4.0
pip-upgrader==1.4.15
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pydantic==2.8.2
pydantic_core==2.23.0
pygame==2.6.0
pyparsing==3.1.2
pyscreenshot==3.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
setuptools==73.0.1
SpeechRecognition==3.10.4
terminaltables==3.1.10
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
